feat: Initialize project with Expo and add essential screens
- Created package.json with dependencies for React Navigation, Expo, and testing libraries.
- Added AboutScreen, ContactScreen, HomeScreen, NewsScreen, PhotoCaptureScreen, RegistrationScreen, WeatherScreen, and ExplicitPermissionScreen.
- Implemented form validation utility functions for registration.
- Removed outdated test files related to registration validation.
- Configured TypeScript with tsconfig.json.
- Added README for screenshots directory.
