+++ package.json
@@ -1,0 +1,41 @@
+{
+  "name": "react-expo-hands-on",
+  "version": "1.0.0",
+  "main": "index.ts",
+  "scripts": {
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web",
+    "test": "jest --config jest.config.js --verbose"
+  },
+  "dependencies": {
+    "@react-navigation/native": "^7.1.16",
+    "@react-navigation/native-stack": "^7.3.23",
+    "expo": "~53.0.20",
+    "expo-camera": "~16.1.11",
+    "expo-linking": "~7.1.7",
+    "expo-media-library": "~17.1.7",
+    "expo-status-bar": "~2.2.3",
+    "react": "19.0.0",
+    "react-native": "0.79.5",
+    "react-native-gesture-handler": "~2.24.0",
+    "react-native-reanimated": "~3.17.4",
+    "react-native-safe-area-context": "5.4.0",
+    "react-native-screens": "~4.11.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.28.0",
+    "@babel/preset-react": "^7.27.1",
+    "@babel/preset-typescript": "^7.27.1",
+    "@testing-library/react-native": "^13.2.0",
+    "@types/jest": "^30.0.0",
+    "@types/react": "~19.0.10",
+    "babel-jest": "^30.0.5",
+    "jest": "^30.0.5",
+    "typescript": "~5.8.3",
+    "react-test-renderer": "19.0.0"
+  },
+  "private": true
+}
