--- react-expo-hands-on/babel.config.js
@@ -1,4 +1,0 @@
-// babel.config.js
-module.exports = {
-  presets: ['babel-preset-expo'],
-};
