--- react-expo-hands-on/jest.config.js
@@ -1,8 +1,0 @@
-// jest.pure.config.js
-module.exports = {
-  testMatch: ['**/__tests__/**/*.test.ts?(x)'], // only pure logic tests
-  transform: {
-    '^.+\\.[jt]sx?$': 'babel-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-};
