--- react-expo-hands-on/screens/AboutScreen.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-const AboutScreen = () => {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.text}>This is the About Screen</Text>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 20,
-  },
-});
-
-export default AboutScreen;
