+++ screens/RegistrationScreen.tsx
@@ -1,0 +1,77 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, Button, StyleSheet } from 'react-native';
+
+const RegistrationScreen = () => {
+  const [name, setName] = useState('');
+  const [email, setEmail] = useState('');
+  const [error, setError] = useState('');
+
+  const handleRegister = () => {
+    if (!name || !email) {
+      setError('All fields are required!');
+      return;
+    }
+    if (!email.includes('@')) {
+      setError('Enter a valid email!');
+      return;
+    }
+    setError('');
+    // Here, you would submit the registration details to an API or backend
+    alert('Registration successful!');
+    setName('');
+    setEmail('');
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>User Registration</Text>
+      {error ? <Text style={styles.error}>{error}</Text> : null}
+      <TextInput
+        style={styles.input}
+        placeholder="Name"
+        value={name}
+        onChangeText={setName}
+      />
+      <TextInput
+        style={styles.input}
+        placeholder="Email"
+        keyboardType="email-address"
+        value={email}
+        onChangeText={setEmail}
+      />
+      <Button title="Register" onPress={handleRegister} />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 16,
+    backgroundColor: '#fff',
+  },
+  title: {
+    fontSize: 22,
+    marginBottom: 20,
+    fontWeight: 'bold',
+  },
+  input: {
+    width: '100%',
+    maxWidth: 300,
+    padding: 10,
+    marginBottom: 16,
+    borderWidth: 1,
+    borderColor: '#ccc',
+    borderRadius: 6,
+    fontSize: 16,
+    backgroundColor: '#f9f9f9',
+  },
+  error: {
+    color: 'red',
+    marginBottom: 10,
+  },
+});
+
+export default RegistrationScreen;
