+++ screenshots/README.md
@@ -1,0 +1,1 @@
+## This directory contains screenshots of the different tasks mentioned in the hands-on document
